+++ src/O4_DSF_Utils.py
@@ -14,7 +14,7 @@ import numpy
 from PIL import Image, ImageDraw
 
 from . import airport_data as APT_SRC
-from . import O4_File_Names as FNAMES
+from . import filenames as FNAMES
 from . import O4_Geo_Utils as GEO
 from . import O4_Mask_Utils as MASK
 from . import O4_UI_Utils as UI
