+++ src/O4_GUI_Utils.py
@@ -36,7 +36,7 @@ from . import airport_data as APT_SRC
 from . import common
 from . import config as CFG
 from . import O4_DSF_Utils as DSF
-from . import O4_File_Names as FNAMES
+from . import filenames as FNAMES
 from . import O4_Geo_Utils as GEO
 from . import O4_Imagery_Utils as IMG
 from . import O4_Mask_Utils as MASK
