+++ src/O4_Imagery_Utils.py
@@ -12,7 +12,7 @@ import requests
 from PIL import Image, ImageEnhance, ImageFilter, ImageOps
 
 Image.MAX_IMAGE_PIXELS = 1000000000  # Not a decompression bomb attack!
-from . import O4_File_Names as FNAMES
+from . import filenames as FNAMES
 from . import O4_Geo_Utils as GEO
 from . import O4_UI_Utils as UI
