+++ src/O4_Mesh_Utils.py
@@ -9,7 +9,7 @@ import numpy
 import requests
 
 from . import O4_DEM_Utils as DEM
-from . import O4_File_Names as FNAMES
+from . import filenames as FNAMES
 from . import O4_Geo_Utils as GEO
 from . import O4_OSM_Utils as OSM
 from . import O4_UI_Utils as UI
