+++ src/airport.py
@@ -7,7 +7,7 @@ from rtree import index
 from shapely import affinity, geometry, ops
 
 from . import O4_DEM_Utils as DEM
-from . import O4_File_Names as FNAMES
+from . import filenames as FNAMES
 from . import O4_Geo_Utils as GEO
 from . import O4_UI_Utils as UI
 from . import O4_Vector_Utils as VECT
