+++ src/airport_data.py
@@ -15,7 +15,7 @@ import shapely.ops
 import shapely.prepared
 
 from . import config as CFG
-from . import O4_File_Names as FNAMES
+from . import filenames as FNAMES
 from . import O4_Geo_Utils as GEO
 from .common import IcaoCode
